deleted unnecessary components and renamed util folder as assets
