+++ src/pages/productItem.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 import _ from 'lodash';
-import { createCart, addToCart, removeFromCart } from '../utils/shopifyUtils';
+import { createCart, addToCart, removeFromCart } from '../assets/shopifyUtils';
 
 export default class ProductItem extends React.Component {
   constructor(props) {
